new logaddexp, float vs double transition

+++ include/imm/float.h
@@ -1,6 +1,6 @@
 #ifndef IMM_FLOAT_H
 #define IMM_FLOAT_H
 
-typedef double imm_float;
+typedef float imm_float;
 
 #endif
